requirements.txt
telethon==1.33.1
python-dotenv==1.0.1
cryptg==0.4.0
pillow==10.2.0
aiohttp==3.9.3